requirements.txt
python-dotenv
python3-discogs-client
fastapi[all]
jinja2
jinja2-fragments
python-multipart
pytest
pytailwindcss
tinydb